refactor: add day component
